quiz_app/requirements.txt
Django
django-extensions
django-model-utils
django-localflavor
djangorestframework
django-rest-auth
coreapi
psycopg2
django-cors-headers